+++ libs/steam-utility/Commands/ICommand.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace SteamUtility.Commands
+{
+    public interface ICommand
+    {
+        void Execute(string[] args);
+    }
+}