+++ changelog.md
@@ -1,5 +1,6 @@
 [![Downloads](https://img.shields.io/github/downloads/probablyraging/steam-game-idler/1.5.26/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/probablyraging/steam-game-idler/releases/download/1.5.26/Steam.Game.Idler_1.5.26_x64_en-US.msi)
 
 ## Changelog
-- Added `Free game notifications` to `settings > general`
-  - Receive native Windows notifications when free Steam games are available to add to your account
+- Fixed some UI blocking issues when running specific Tauri commands
+- Improved handling for unlocking/locking achievements, especially when done in bulk
+- Refactored the `SteamUtility` lib